+++ packages/server/api/src/app/app.ts
@@ -84,7 +84,7 @@ import { userModule } from './user/user.module'
 import { webhookModule } from './webhooks/webhook-module'
 import { websocketService } from './websockets/websockets.service'
 import { flowQueueConsumer } from './workers/flow-worker/consumer/flow-queue-consumer'
-import { engineResponseWatcher } from './workers/flow-worker/engine-response-watcher'
+import { webhookResponseWatcher } from './workers/flow-worker/webhook-response-watcher'
 import { flowWorkerModule } from './workers/flow-worker/flow-worker-module'
 import { setupBullMQBoard } from './workers/flow-worker/queues/redis/redis-bullboard'
 import {
@@ -367,7 +367,7 @@ export const setupApp = async (): Promise<FastifyInstance> => {
     app.addHook('onClose', async () => {
         await flowQueueConsumer.close()
         await systemJobsSchedule.close()
-        await engineResponseWatcher.shutdown()
+        await webhookResponseWatcher.shutdown()
     })
 
     return app
