+++ packages/server/api/src/app/flows/flow-run/flow-run-module.ts
@@ -1,10 +1,10 @@
 import { FastifyPluginAsync } from 'fastify'
 import { entitiesMustBeOwnedByCurrentProject } from '../../authentication/authorization'
-import { engineResponseWatcher } from '../../workers/flow-worker/engine-response-watcher'
+import { webhookResponseWatcher } from '../../workers/flow-worker/webhook-response-watcher'
 import { flowRunController as controller } from './flow-run-controller'
 
 export const flowRunModule: FastifyPluginAsync = async (app) => {
     app.addHook('preSerialization', entitiesMustBeOwnedByCurrentProject)
     await app.register(controller, { prefix: '/v1/flow-runs' })
-    await engineResponseWatcher.init()
+    await webhookResponseWatcher.init()
 }
