+++ packages/server/api/src/app/flows/flow-run/flow-run-service.ts
@@ -8,7 +8,7 @@ import { buildPaginator } from '../../helper/pagination/build-paginator'
 import { paginationHelper } from '../../helper/pagination/pagination-utils'
 import { Order } from '../../helper/pagination/paginator'
 import { telemetry } from '../../helper/telemetry.utils'
-import { engineResponseWatcher } from '../../workers/flow-worker/engine-response-watcher'
+import { webhookResponseWatcher } from '../../workers/flow-worker/webhook-response-watcher'
 import { flowService } from '../flow/flow.service'
 import { FlowRunEntity } from './flow-run-entity'
 import { flowRunHooks } from './flow-run-hooks'
@@ -84,7 +84,7 @@ async function updateFlowRunToLatestFlowVersionId(
 }
 
 function returnHandlerId(pauseMetadata: PauseMetadata | undefined, requestId: string | undefined): string {
-    const handlerId = engineResponseWatcher.getHandlerId()
+    const handlerId = webhookResponseWatcher.getHandlerId()
     if (isNil(pauseMetadata)) {
         return handlerId
     }
@@ -302,7 +302,7 @@ export const flowRunService = {
             payload,
             environment: RunEnvironment.TESTING,
             executionType: ExecutionType.BEGIN,
-            synchronousHandlerId: engineResponseWatcher.getHandlerId(),
+            synchronousHandlerId: webhookResponseWatcher.getHandlerId(),
             progressUpdateType: ProgressUpdateType.TEST_FLOW,
         })
     },
