+++ packages/server/api/src/app/flows/flow/flow-worker.controller.ts
@@ -1,7 +1,7 @@
 import { FastifyPluginAsyncTypebox, Type } from '@fastify/type-provider-typebox'
 import { StatusCodes } from 'http-status-codes'
 import { entitiesMustBeOwnedByCurrentProject } from '../../authentication/authorization'
-import { EngineHttpResponse, engineResponseWatcher } from '../../workers/flow-worker/engine-response-watcher'
+import { EngineHttpResponse, webhookResponseWatcher } from '../../workers/flow-worker/webhook-response-watcher'
 import { flowRunService } from '../flow-run/flow-run-service'
 import { flowVersionService } from '../flow-version/flow-version.service'
 import { flowService } from './flow.service'
@@ -30,7 +30,7 @@ export const flowWorkerController: FastifyPluginAsyncTypebox = async (fastify) =
     fastify.post('/update-run', UpdateStepProgress, async (request) => {
         const { runId, workerHandlerId, runDetails, progressUpdateType } = request.body
         if (progressUpdateType === ProgressUpdateType.WEBHOOK_RESPONSE && workerHandlerId) {
-            await engineResponseWatcher.publish(
+            await webhookResponseWatcher.publish(
                 runId,
                 workerHandlerId,
                 await getFlowResponse(runDetails),
