+++ packages/server/api/src/app/workers/flow-worker/consumer/webook-consumer.ts
@@ -1,7 +1,7 @@
 import { StatusCodes } from 'http-status-codes'
 import { flowService } from '../../../flows/flow/flow.service'
 import { webhookService } from '../../../webhooks/webhook-service'
-import { EngineHttpResponse, engineResponseWatcher } from '../engine-response-watcher'
+import { EngineHttpResponse, webhookResponseWatcher } from '../webhook-response-watcher'
 import { FlowStatus, isNil } from '@activepieces/shared'
 import { WebhookJobData } from 'server-worker'
 
@@ -47,7 +47,7 @@ export const webhookConsumer = {
         }
         const runs = await webhookService.callback({
             flow,
-            synchronousHandlerId: engineResponseWatcher.getHandlerId(),
+            synchronousHandlerId: webhookResponseWatcher.getHandlerId(),
             payload,
         })
         if (isNil(runs) || runs.length === 0 || isNil(runs[0])) {
@@ -60,7 +60,7 @@ export const webhookConsumer = {
         }
         if (!isNil(data.synchronousHandlerId)) {
             const firstRun = runs[0]
-            const response = await engineResponseWatcher.oneTimeListener(firstRun.id, true)
+            const response = await webhookResponseWatcher.oneTimeListener(firstRun.id, true)
             await stopAndReply(data, response)
         }
     },
@@ -70,7 +70,7 @@ export const webhookConsumer = {
 async function stopAndReply(data: WebhookJobData, response: EngineHttpResponse): Promise<void> {
     const { requestId, synchronousHandlerId } = data
     if (!isNil(synchronousHandlerId)) {
-        await engineResponseWatcher.publish(
+        await webhookResponseWatcher.publish(
             requestId,
             synchronousHandlerId,
             response,
