+++ packages/server/api/test/integration/ce/flows/flow-consume.test.ts
@@ -1,8 +1,8 @@
 import { FastifyInstance } from 'fastify'
 import { setupApp } from '../../../../src/app/app'
 import { databaseConnection } from '../../../../src/app/database/database-connection'
-import { engineResponseWatcher } from '../../../../src/app/workers/flow-worker/engine-response-watcher'
 import { flowWorker } from '../../../../src/app/workers/flow-worker/flow-worker'
+import { webhookResponseWatcher } from '../../../../src/app/workers/flow-worker/webhook-response-watcher'
 import {
     createMockFlow,
     createMockFlowRun,
@@ -137,7 +137,7 @@ describe('flow execution', () => {
             environment: RunEnvironment.PRODUCTION,
             runId: mockFlowRun.id,
             payload: {},
-            synchronousHandlerId: engineResponseWatcher.getHandlerId(),
+            synchronousHandlerId: webhookResponseWatcher.getHandlerId(),
             progressUpdateType: ProgressUpdateType.NONE,
             executionType: ExecutionType.BEGIN,
         })
